splitting up button view to have a separate component on seating chart page, since it's getting too different from the ones on class pages too manage
