+++ client/src/components/TeacherView/ButtonView.jsx
@@ -2,7 +2,6 @@ import React from "react";
 
 // This component is for the buttons that are sandwich/notepaper color
 // // mainly the two buttons in manage class list and classroom, 
-// // as well as in the edit seating chart for student roster and classroom furniture
 
 const ButtonView = ({
   buttonText,
@@ -17,15 +16,15 @@ const ButtonView = ({
     <>
       <div className={`${buttonSize !== "small" ? "mx-2" : ""}`}>
         <button
-          className={`text-[Poppins] rounded-xl flex items-center ${buttonSize === "small" ? "border-[2px] border-sandwich flex-row-reverse md:flex-col-reverse h-[50px] md:h-24 w-[140px] xs:w-[150px] md:w-20 text-[13px] justify-center" : "text-[16px] border-[4px] border-sandwich h-20 w-[270px] justify-around" }   ${
+          className={`text-[Poppins] rounded-xl flex items-center border-[2px] border-sandwich flex-col-reverse h-24 w-20 text-[13px] justify-center ${
             isSelected ? "underline font-[700] bg-sandwich" : "bg-notebookPaper"
           }`}
           onClick={handleClick}
         >
-          <h4 className="pl-2 md:pl-0 md:pt-2">{buttonText}</h4>
+          <h4 className="md:pl-0 pt-2">{buttonText}</h4>
           <div className="h-6 flex items-center">
           <img
-          className="w-10 pl-2 xs:w-6"
+          className="w-6"
             src={!isSelected ? defaultBtnImage : btnImageWhenOpen}
             alt={buttonText}
           />
