+++ client/src/pages/teacher/EditSeatingChart.jsx
@@ -20,7 +20,7 @@ import FurnitureImg from "../../images/Desk.png";
 import openRosterImg from "../../images/ThreePplLight.png";
 import openFurnitureImg from "../../images/DeskImgLight.png";
 import MsgModal from "../../components/SeatingChart/MsgModal";
-import ButtonView from "../../components/ButtonView";
+import SeatingChartButton from "../../components/TeacherView/SeatingChartButton";
 import BtnRainbow from "../../components/BtnRainbow";
 import Nav from "../../components/Navbar/Nav";
 import withAuth from "../../hoc/withAuth";
@@ -256,7 +256,7 @@ const EditSeatingChart = () => {
             </div>
             {/* Room View & List Buttons */}
             <div className="flex md:hidden justify-around md:justify-between gap-2 md:gap-4 items-center mt-5 bg-notebookPaper">
-                <ButtonView
+                <SeatingChartButton
                   buttonText="Student Roster"
                   defaultBtnImage={RosterImg}
                   btnImageWhenOpen={openRosterImg}
@@ -267,7 +267,7 @@ const EditSeatingChart = () => {
                   isSelected={showStudentRosterModal}
                   buttonSize="small"
                 />
-                  <ButtonView
+                  <SeatingChartButton
                     buttonText="Classroom Objects"
                     defaultBtnImage={FurnitureImg}
                     btnImageWhenOpen={openFurnitureImg}
@@ -364,7 +364,7 @@ const EditSeatingChart = () => {
           <div className="flex flex-col gap-4 md:gap-0 md:flex-row w-full justify-center items-center md:mt-10">
             {/* Open Choose Students Modal */}
             <div className="hidden md:flex flex-col md:flex-row gap-4 items-center justify-center">
-            <ButtonView
+            <SeatingChartButton
               buttonText="Student Roster"
               defaultBtnImage={RosterImg}
               btnImageWhenOpen={openRosterImg}
@@ -378,7 +378,7 @@ const EditSeatingChart = () => {
 
             {/* Open Choose Furniture Modal */}
 
-            <ButtonView
+            <SeatingChartButton
               buttonText="Classroom Objects"
               defaultBtnImage={FurnitureImg}
               btnImageWhenOpen={openFurnitureImg}
